requirements.txt
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.9
dj-database-url==0.4.1
Django==1.9
django-filter==0.13.0
djangorestframework==3.3.3
ipython==4.2.0
ipython-genutils==0.1.0
Markdown==2.6.6
pexpect==4.1.0
pickleshare==0.7.2
ptyprocess==0.5.1
pytz==2016.4
simplegeneric==0.8.1
traitlets==4.2.1
